requirements.txt
asgiref==3.7.2
Django==4.2.5
django-cleanup==8.0.0
django-modeltranslation==0.18.11
django-tinymce==3.6.1
Pillow==10.0.1
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
whitenoise==6.5.0
